backend/requirements.txt
alembic==1.17.2
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
black==25.11.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
colorama==0.4.6
distro==1.9.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
Mako==1.3.10
MarkupSafe==3.0.3
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
pytokens==0.3.0
requests==2.32.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
